% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_firms.R
\name{fetch_firms}
\alias{fetch_firms}
\title{Fetch FIRMS fire data from NASA's FIRMS API}
\usage{
fetch_firms(
  api_key,
  region_sf,
  start_date,
  end_date,
  dataset = c("VIIRS_SNPP_NRT", "MODIS_NRT"),
  confidence_level = NULL
)
}
\arguments{
\item{api_key}{Character. Your NASA API key. You can register for one at \url{https://firms.modaps.eosdis.nasa.gov/api/map_key}.}

\item{region_sf}{An \code{sf} object representing the area of interest. It will be reprojected to WGS 84 if needed.}

\item{start_date}{Start date of the time range to fetch fire data (in "YYYY-MM-DD" format).}

\item{end_date}{End date of the time range to fetch fire data (in "YYYY-MM-DD" format).}

\item{dataset}{Character. Choose between \code{"VIIRS_SNPP_NRT"} or \code{"MODIS_NRT"}. Defaults to \code{"VIIRS_SNPP_NRT"}.}

\item{confidence_level}{Optional. Filter for confidence level using \code{"l"}, \code{"n"}, \code{"h"}.
MODIS confidence level (numeric, 0-100) is automatically converted to these characters.}
}
\value{
An \code{sf} object containing fire detections within the selected time range and area.

#' @importFrom sf st_crs st_transform st_bbox
}
\description{
This function retrieves fires detected by NASA's FIRMS (Fire Information for Resource Management System) via their API
For more information visit \url{https://firms.modaps.eosdis.nasa.gov/api/}.
Data will be based on a specified region, time range, and dataset.
If the request exceeds 10 days, it will automatically split the request into multiple calls to respect the API limits.
}
