% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_season.R
\name{fire_seasons}
\alias{fire_seasons}
\title{Classify fires by custom-defined fire seasons}
\usage{
fire_seasons(firms_sf, fire_seasons)
}
\arguments{
\item{firms_sf}{An sf object with an \code{acq_date} column.}

\item{fire_seasons}{A named list, where each element is a vector of month numbers (1–12).}
}
\value{
The input \code{sf} object with an added \code{fire_season} column.
}
\description{
Adds a \code{fire_season} column based on user-defined month groupings.
}
