% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_firms.R
\name{plot_firms}
\alias{plot_firms}
\title{Plot raw FIRMS fire detections with brightness and FRP}
\usage{
plot_firms(firms_sf, base_map = NULL)
}
\arguments{
\item{firms_sf}{An \code{sf} object containing FIRMS fire detections. Must include a \code{brightness} or \code{bright_ti4} column,
and an \code{frp} column for Fire Radiative Power.}

\item{base_map}{Optional. An \code{sf} object to use as the base map. If \code{NULL}, a cropped country-level base map from \code{rnaturalearth} will be used.}
}
\value{
A \code{ggplot} object displaying fire points with brightness and FRP on a zoomed map.
}
\description{
This function visualizes FIRMS fire detections (from VIIRS or MODIS) on a base map.
Fire points are color-coded by brightness (\code{brightness} or \code{bright_ti4}, depending on the dataset)
and scaled in size by Fire Radiative Power (\code{frp}).
}
