% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncategorised_fires.R
\name{uncategorized_fires}
\alias{uncategorized_fires}
\title{Identify uncategorized fire detections}
\usage{
uncategorized_fires(firms_sf, classified_firms_list)
}
\arguments{
\item{firms_sf}{An \code{sf} object containing the original FIRMS fire detections.}

\item{classified_firms_list}{A list of \code{sf} objects, each containing a subset of \code{firms_sf} with fire classifications (e.g., from \code{fetch_osm()}).}
}
\value{
An \code{sf} object with uncategorized fire points. A \code{fire_type} column is added and set to \code{"unknown"}.
}
\description{
This function compares a raw FIRMS fire dataset with a list of previously classified fire points.
It returns only those fire detections that were not assigned a classification (i.e., did not appear in any of the classified datasets).
}
