% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fire_season.R
\name{plot_fire_seasons}
\alias{plot_fire_seasons}
\title{Plot fire season statistics and spatial distribution}
\usage{
plot_fire_seasons(firms_sf, base_map = NULL)
}
\arguments{
\item{firms_sf}{An sf object with columns \code{fire_type} and \code{fire_season}}

\item{base_map}{Optional. Background map (sf object or file path). Default is Natural Earth.}

\item{fire_season_order}{Optional. A character vector with the desired fire season order (e.g., names from classify_fire_seasons)}
}
\value{
A list with two ggplot objects: $bar_plot and $map_plot
}
\description{
Plot fire season statistics and spatial distribution
}
